+++ src/mjlab/tasks/path_tracking/config/nugus/__init__.py
@@ -1,5 +1,5 @@
+from mjlab.tasks.path_tracking.rl import PathTrackingOnPolicyRunner
 from mjlab.tasks.registry import register_mjlab_task
-from mjlab.tasks.velocity.rl import VelocityOnPolicyRunner
 
 from .env_cfgs import (
   nubots_nugus_path_flat_env_cfg,
@@ -12,7 +12,7 @@ register_mjlab_task(
   env_cfg=nubots_nugus_path_rough_env_cfg(),
   play_env_cfg=nubots_nugus_path_rough_env_cfg(play=True),
   rl_cfg=nubots_nugus_path_ppo_runner_cfg(),
-  runner_cls=VelocityOnPolicyRunner,
+  runner_cls=PathTrackingOnPolicyRunner,
 )
 
 register_mjlab_task(
@@ -20,5 +20,5 @@ register_mjlab_task(
   env_cfg=nubots_nugus_path_flat_env_cfg(),
   play_env_cfg=nubots_nugus_path_flat_env_cfg(play=True),
   rl_cfg=nubots_nugus_path_ppo_runner_cfg(),
-  runner_cls=VelocityOnPolicyRunner,
+  runner_cls=PathTrackingOnPolicyRunner,
 )